+++ src/app/modules/core/models/layers/layers-provider.ts
@@ -25,6 +25,4 @@ export interface LayersProvider {
    */
   getUiControlsComponentType(model: Layer): Type<LayerAware>;
 
-  // TODO add a getter for the layer code? (would allow to get rid of SupportedLayers)
-
 }