+++ src/app/modules/core/services/factories/aggregate-layer-factory.service.ts
@@ -5,12 +5,16 @@ import { Layer } from '#core/models/layers/layer';
 
 /**
  * Factory object for AggregateLayer instances.
+ *
  */
 @Injectable()
 export class AggregateLayerFactoryService implements LayerFactory {
 
   public buildRenderableLayer(model: Layer): AggregateLayer {
-    return new AggregateLayer(model);
+    if (model) {
+      return new AggregateLayer(model);
+    }
+    return undefined;
   }
 
 }